phase-2-mcp-server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "second-brain-mcp-server"
version = "0.1.0"
description = "MCP server connecting Claude.ai to the Second Brain vault."
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.2.0",
    "fastmcp>=0.4.0",
    "qdrant-client[fastembed]>=1.12.0",
    "voyageai>=0.3.0",
    "python-frontmatter>=1.1.0",
    "rapidfuzz>=3.9.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.1.0",
    "uvicorn[standard]>=0.30.0",
    "starlette>=0.40.0",
    "fastapi>=0.110.0",
    "python-slugify>=8.0.0",
    "httpx>=0.27.0",
    "mistune>=3.0.2",
    "cryptography>=42.0.0",
    "google-auth>=2.30.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.130.0",
    "tenacity>=8.3.0",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
mcp-server = "mcp_server.server:main"
gtasks-auth = "mcp_server.gtasks_cli:auth"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
